Extract common Cassandra functionality into a separate project.
The idea is that there will be a new project later, called
`snapshot-cassandra` that will reuse the exctracted functionality
without depending on `eventual-cassandra`.

This may also simplify reading of the code as Cassandra and journal
functionality will be clearly separated.

The packages are not renamed, for now, to keep backwards compatibility
with the downstream projects.