Add SnapshotStoreFlat interface.
This is intended to be a main entry point to a new snapshotter.

The main considerations were to keep it recognizable and easy to use
from a Akka Persistence Plugin, while having no dependencies on Akka
itself.

I also tried to make it look similar to `ReplicatedJournalFlat` and
`EventualJournal` for sake of uniformity of the codebase.