Restore binary backwards compatibility of SchemaConfig.
The idea is to restore binary backwards compatibility of `SchemaConfig`.

`SchemaConfig.Keyspace` is restored for that purposes and the method is
added to convert it to `KeyspaceConfig`.

The class itself is marked as deprecated and is planned to be removed in
a next non-binary-compatible release.

This is a smaller PR split from #586 for easier review.