Restore binary backwards compatibility for CreateTables and CassandraSync.
**Details**

There are now two copies of these classes with a same name. One in
`cassandra` package and one in `eventual.cassandra` package.  It was a
bit easier to move both CreateTables and CassandraSync at once, because
they are related.

Also improved CreateTables unit tests to make them more readable and
easier to maintain.

**Reasoning**

See the reasoning behind the approach in the following PR: #594