Restore binary backwards compatibility for SettingsStatements and SettingsCassandra (#602)
**Details**

There are now two copies of these classes with a same name. One in
cassandra package and one in eventual.cassandra package. It was a bit
easier to move both SettingsCassandra and SettingsStatements at once,
because they are related.

Also improved SettingsCassandra unit tests to make them more
readable and easier to maintain. The integration tests were updated
to test both a new class and a compatibility wrapper.

**Reasoning**

See the reasoning behind the approach in the following PR: #594