Move `MigrateSchema` from `eventual.cassandra` to `cassandra`.
This is a new class, common to the eventual Cassandra journal storage
and to a new snapshotter, so it does not make sense to keep it in the
old path.

To remind: we only kept the classes at this path in `cassandra` sbt
subproject for sake of binary compatibility with a previous version.