Fixed flakiness of DamperTest again.
The original fix made in #555 fixed the original source of flakiness,
but introduced a different one by an assertion added to a wrong place.
It would check if `delayOf` was called only once, while starting the
threads, when cancelation was already started.

The fiber cancelation would also call `delayOf` in a background, and the
test would fail from time to time depending on how quickly this call
propagates to a `ref` used in a test.