Add ReplicatedOffsetNotifier callback
Current recovery logic doesn't scale when the volume of events is large,
because clients have to consume all the events on all the app nodes.
Additionally, current logic creates Kafka consumers on-demand,
which is a heavy operation. This consumer creation happens in spikes,
when replicator is running late, and affects client applications stability.

Introducing a side channel for notifications about replication progress
will allow us in the future to have an alternative implementation for recovery
without these flaws. I.e. notifications go to a separate lean Kafka topic
which is read by all client nodes. On recovery:
- a mark is written
- wait until mark offset confirmed replicated in the notification topic
  (with potential fallback to polling of the pointer table in Cassandra)
- recover from Cassandra only