Prepare for Scala 3 support - remove scodec upcast usage (#757)
In the Events type scodec usage didn't compile with Scala3: upcast
method didn't work if a type is wrapped in Option. To fix that, the
Events codec was refactored to use the discriminated method.

Also:
- refactored serde tests against file examples to make them more robust
and readable
- fixed EventsToBytesSpec not verifying exact byte outputs for the
current format which allowed format errors to slip through
- deprecated PayloadType.ext because it was used only in test code and
didn't seem to belong in the main model code